int -> float

+++ readFile.py
@@ -1,5 +1,5 @@
 def readFile(filename):
-    return [int(i) for i in open(filename, 'r').read().strip().split('\n')]
+    return [float(i) for i in open(filename, 'r').read().strip().split('\n')]
 
 numbers = readFile('numbers.txt')
